Refactor of the submission utility
The submission utility in application has become a mess since
it's creation due a limited view of it's usage at conception
and extending it's functionality over time.

This overhaul of the utility accomplishes the following:
- Moves from passing call back functions to using async utility functions
- Moves pipeline related utility functions into pipeline-editor (Part of #494)
- Update all calls to metadata end points to use the metadata service. Fixes #271
- Move ErrorDialogContent into it's own file. First step towards #403
- Update function, class, and file names to match actual functionality
- Modularized HTTP_STATUS success code verification. Fixes #232

Fixes #400
